Merged PR 52293: Create new module for Page Builder
## What's being changed

We have created a new module for Adobe's Page Builder extension.

## Why it's being changed

We want to move the Dotdigital Page Builder content type into a separate module, since as of v2.4.0 this is included with Magento Open Source. This will allow us to also make modifications to ensure the module is compatible with Hyvä.

## How to review / test this change

### Page Builder
- Run `setup:upgrade`
- Run `di:compile`
- Observe the page builder feature in the content sections of a page and interactions. See no errors.
- Add 2 forms to a page: 1 popover and 1 embedded, ensuring they both have the email field
- Observe the page and see no errors
- Fill in the email field forms and observe an identify call in the networks tab with the related email.
- Submit the form and ensure it behaves as expected.
- Observe "Dotdigital Page Builder for Magento 2: v. 1.0.0" on the dotdigital reports dashboard.
- Run `vendor/bin/mftf run:group DotDigitalPageBuilder` see a green build

Related work items: #153783